requirements.txt
sentence-transformers
faiss-cpu
httpx
pydantic
openai
tavily-python
